--- Moisha-Front/src/app/components/side-bar/side-bar.component.html
@@ -1,3 +1,0 @@
-<p>
-  side-bar works!
-</p>
