+++ Moisha-Front/src/app/core/interest.service.ts
@@ -41,5 +41,7 @@ export class InterestService {
   createInterest(payload) {
     return this.http.post('/interest/create/', payload);
   }
-
+  getUserInterests(): Observable<Interest[]> {
+    return this.http.get<Interest[]>('/interest/user').pipe(tap((result) => console.log(result)))
+  }
 }
