+++ Moisha-Front/src/app/feed/home/home.component.html
@@ -1,4 +1,5 @@
 <app-navbar [name]="auth?.user?.name" (search) = "searchInterest($event)"></app-navbar>
+<app-side-bar></app-side-bar>
 <section class="feed-app">
 <router-outlet></router-outlet>
 </section>
