+++ Moisha-Front/src/app/shared/side-bar/side-bar.component.css
@@ -1,0 +1,25 @@
+.container{
+  position: fixed;
+  width: 10%;
+  height: 100%;
+  background-color: #007bff;
+}
+.gab{
+  height: 20%;
+}
+.list-group{
+  overflow-y: scroll;
+  max-height: 60%;
+}
+.badge{
+  text-align: center;
+}
+.list-group-item{
+  padding: 10px 0px;
+  margin: 5px;
+  background-color: aliceblue;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  border-bottom-left-radius: 10px;
+  border-bottom-right-radius: 10px;
+}
