+++ Moisha-Front/src/app/shared/side-bar/side-bar.component.html
@@ -1,0 +1,9 @@
+<div class="container">
+  <div class="gab"></div>
+  <ul class="interests list-group">
+    <a *ngFor="let interest of interests" class="list-group-item"
+       routerLink="/interest/{{interest.id}}">
+      <span class="badge">{{interest.name}}</span>
+    </a>
+  </ul>
+</div>
