+++ Moisha_Back/interest/urls.py
@@ -4,4 +4,5 @@ from interest import views
 
 urlpatterns = [
     path('create/', views.createInterest),
+    path('user/', views.getInterestsByUser),
 ]
