+++ Moisha_Back/interest/views.py
@@ -16,3 +16,10 @@ def createInterest(request):
         return Response(serializer.data)
     print(serializer.errors)
     return Response(data=serializer.errors, status= status.HTTP_400_BAD_REQUEST)
+
+@api_view(['GET'])
+def getInterestsByUser(request):
+    user = request.user
+    if user.is_anonymous:
+        return Response('Anonymous user is not allowed', status=status.HTTP_400_BAD_REQUEST)
+    return Response(data=InterestSerializer(user.interests, many=True).data, status=status.HTTP_200_OK)
